feat(Post): 게시글에 대해 인가 적용

+++ src/posts/service/post_service.py
@@ -1,4 +1,5 @@
 from fastapi import HTTPException
+from starlette import status
 
 from posts.model.post import Post
 from posts.repository.post_repository import PostRepository
@@ -18,17 +19,20 @@ class PostService:
             raise HTTPException(status_code=404, detail="Post not found")
         return post
 
-    def create_post(self, request: CreatePostRequest) -> Post:
+    def create_post(self, request: CreatePostRequest, member_id: int) -> Post:
         post = Post(
             title=request.title,
             content=request.content,
             article_image=request.article_image_url,
-            author_id=request.author_id
+            author_id=member_id
         )
         return self.post_repository.save(post)
 
-    def update_post(self, post_id: int, request: UpdatePostRequest) -> Post:
+    def update_post(self, post_id: int, request: UpdatePostRequest, member_id: int) -> Post:
         post = self.get_post(post_id)
+
+        if post.author_id != member_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
 
         if request.title is not None:
             post.title = request.title
@@ -39,6 +43,8 @@ class PostService:
 
         return self.post_repository.save(post)
 
-    def delete_post(self, post_id: int) -> None:
+    def delete_post(self, post_id: int, member_id: int) -> None:
         post = self.get_post(post_id)
+        if post.author_id != member_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
         self.post_repository.delete(post)
